requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
requests==2.31.0

# Support de l'envoi de fichiers (Crucial pour /v1/asr)
python-multipart==0.0.6

# ASR (Reconnaissance vocale)
faster-whisper==1.0.3

# Dépendances NVIDIA CUDA (Version 12 pour correspondre à ton erreur cublas64_12)
nvidia-cublas-cu12==12.1.3.1
nvidia-cudnn-cu12==8.9.2.26

# Gestion des chemins (Optionnel mais recommandé pour Python 3.10+)
pathlib==1.0.1
python-dotenv

pymongo[srv]
spacy
networkx